+++ lib/changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - \[Demo\] Add `ComponentCustomizationChipRow` and `ComponentCustomizationChip` composables to display rows of chips in bottom sheets ([#141](https://github.com/Orange-OpenSource/ods-android/issues/141))
+- \[Demo\] Add guideline spacings screen ([#211](https://github.com/Orange-OpenSource/ods-android/issues/211))
 - \[Lib\] Add `OdsListItemScope` and `OdsListItemIconType` enum as well as `Modifier.iconType` and `Modifier.divider` methods to configure `OdsListItem` icon type and divider ([#141](https://github.com/Orange-OpenSource/ods-android/issues/141))
 
 ### Changed